Struct aws_sdk_sesv2::client::fluent_builders::SendEmail

pub struct SendEmail<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> SendEmail<C>

pub async fn send(self) -> Result<SendEmailOutput, SdkError<SendEmailError>> where
C: SmithyConnector,

pub fn from_email_address(self, input: impl Into<String>) -> Self

The email address that you want to use as the "From" address for the email. The address that you specify has to be verified.

pub fn set_from_email_address(self, input: Option<String>) -> Self

pub fn from_email_address_identity_arn(self, input: impl Into<String>) -> Self

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the FromEmailAddress parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use sender@example.com, then you would specify the FromEmailAddressIdentityArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the FromEmailAddress to be sender@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

For Raw emails, the FromEmailAddressIdentityArn value overrides the X-SES-SOURCE-ARN and X-SES-FROM-ARN headers specified in raw email message content.

pub fn set_from_email_address_identity_arn(self, input: Option<String>) -> Self

pub fn destination(self, input: Destination) -> Self

An object that contains the recipients of the email message.

pub fn set_destination(self, input: Option<Destination>) -> Self

pub fn reply_to_addresses(self, inp: impl Into<String>) -> Self

The "Reply-to" email addresses for the message. When the recipient replies to the message, each Reply-to address receives the reply.

pub fn set_reply_to_addresses(self, input: Option<Vec<String>>) -> Self

pub fn feedback_forwarding_email_address(self, input: impl Into<String>) -> Self

The address that you want bounce and complaint notifications to be sent to.

pub fn set_feedback_forwarding_email_address(
self,
input: Option<String>
) -> Self

pub fn feedback_forwarding_email_address_identity_arn(
self,
input: impl Into<String>
) -> Self

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the FeedbackForwardingEmailAddress parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the FeedbackForwardingEmailAddressIdentityArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the FeedbackForwardingEmailAddress to be feedback@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

pub fn set_feedback_forwarding_email_address_identity_arn(
self,
input: Option<String>
) -> Self

pub fn content(self, input: EmailContent) -> Self

An object that contains the body of the message. You can send either a Simple message Raw message or a template Message.

pub fn set_content(self, input: Option<EmailContent>) -> Self

pub fn email_tags(self, inp: impl Into<MessageTag>) -> Self

A list of tags, in the form of name/value pairs, to apply to an email that you send using the SendEmail operation. Tags correspond to characteristics of the email that you define, so that you can publish email sending events.

pub fn set_email_tags(self, input: Option<Vec<MessageTag>>) -> Self

pub fn configuration_set_name(self, input: impl Into<String>) -> Self

The name of the configuration set that you want to use when sending the email.

pub fn set_configuration_set_name(self, input: Option<String>) -> Self

pub fn list_management_options(self, input: ListManagementOptions) -> Self

An object used to specify a list or topic to which an email belongs, which will be used when a contact chooses to unsubscribe.

pub fn set_list_management_options(
self,
input: Option<ListManagementOptions>
) -> Self