Struct aws_sdk_sesv2::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for SimpleEmailService_v2.

This client allows ergonomic access to a SimpleEmailService_v2-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_configuration_set(&self) -> CreateConfigurationSet<C, M, R>

pub fn create_configuration_set_event_destination(
&self
) -> CreateConfigurationSetEventDestination<C, M, R>

pub fn create_contact(&self) -> CreateContact<C, M, R>

pub fn create_contact_list(&self) -> CreateContactList<C, M, R>

pub fn create_custom_verification_email_template(
&self
) -> CreateCustomVerificationEmailTemplate<C, M, R>

pub fn create_dedicated_ip_pool(&self) -> CreateDedicatedIpPool<C, M, R>

pub fn create_deliverability_test_report(
&self
) -> CreateDeliverabilityTestReport<C, M, R>

pub fn create_email_identity(&self) -> CreateEmailIdentity<C, M, R>

pub fn create_email_identity_policy(&self) -> CreateEmailIdentityPolicy<C, M, R>

pub fn create_email_template(&self) -> CreateEmailTemplate<C, M, R>

pub fn create_import_job(&self) -> CreateImportJob<C, M, R>

pub fn delete_configuration_set(&self) -> DeleteConfigurationSet<C, M, R>

pub fn delete_configuration_set_event_destination(
&self
) -> DeleteConfigurationSetEventDestination<C, M, R>

pub fn delete_contact(&self) -> DeleteContact<C, M, R>

pub fn delete_contact_list(&self) -> DeleteContactList<C, M, R>

pub fn delete_custom_verification_email_template(
&self
) -> DeleteCustomVerificationEmailTemplate<C, M, R>

pub fn delete_dedicated_ip_pool(&self) -> DeleteDedicatedIpPool<C, M, R>

pub fn delete_email_identity(&self) -> DeleteEmailIdentity<C, M, R>

pub fn delete_email_identity_policy(&self) -> DeleteEmailIdentityPolicy<C, M, R>

pub fn delete_email_template(&self) -> DeleteEmailTemplate<C, M, R>

pub fn delete_suppressed_destination(
&self
) -> DeleteSuppressedDestination<C, M, R>

pub fn get_account(&self) -> GetAccount<C, M, R>

pub fn get_blacklist_reports(&self) -> GetBlacklistReports<C, M, R>

pub fn get_configuration_set(&self) -> GetConfigurationSet<C, M, R>

pub fn get_configuration_set_event_destinations(
&self
) -> GetConfigurationSetEventDestinations<C, M, R>

pub fn get_contact(&self) -> GetContact<C, M, R>

pub fn get_contact_list(&self) -> GetContactList<C, M, R>

pub fn get_custom_verification_email_template(
&self
) -> GetCustomVerificationEmailTemplate<C, M, R>

pub fn get_dedicated_ip(&self) -> GetDedicatedIp<C, M, R>

pub fn get_dedicated_ips(&self) -> GetDedicatedIps<C, M, R>

pub fn get_deliverability_dashboard_options(
&self
) -> GetDeliverabilityDashboardOptions<C, M, R>

pub fn get_deliverability_test_report(
&self
) -> GetDeliverabilityTestReport<C, M, R>

pub fn get_domain_deliverability_campaign(
&self
) -> GetDomainDeliverabilityCampaign<C, M, R>

pub fn get_domain_statistics_report(&self) -> GetDomainStatisticsReport<C, M, R>

pub fn get_email_identity(&self) -> GetEmailIdentity<C, M, R>

pub fn get_email_identity_policies(&self) -> GetEmailIdentityPolicies<C, M, R>

pub fn get_email_template(&self) -> GetEmailTemplate<C, M, R>

pub fn get_import_job(&self) -> GetImportJob<C, M, R>

pub fn get_suppressed_destination(&self) -> GetSuppressedDestination<C, M, R>

pub fn list_configuration_sets(&self) -> ListConfigurationSets<C, M, R>

pub fn list_contact_lists(&self) -> ListContactLists<C, M, R>

pub fn list_contacts(&self) -> ListContacts<C, M, R>

pub fn list_custom_verification_email_templates(
&self
) -> ListCustomVerificationEmailTemplates<C, M, R>

pub fn list_dedicated_ip_pools(&self) -> ListDedicatedIpPools<C, M, R>

pub fn list_deliverability_test_reports(
&self
) -> ListDeliverabilityTestReports<C, M, R>

pub fn list_domain_deliverability_campaigns(
&self
) -> ListDomainDeliverabilityCampaigns<C, M, R>

pub fn list_email_identities(&self) -> ListEmailIdentities<C, M, R>

pub fn list_email_templates(&self) -> ListEmailTemplates<C, M, R>

pub fn list_import_jobs(&self) -> ListImportJobs<C, M, R>

pub fn list_suppressed_destinations(
&self
) -> ListSuppressedDestinations<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn put_account_dedicated_ip_warmup_attributes(
&self
) -> PutAccountDedicatedIpWarmupAttributes<C, M, R>

pub fn put_account_details(&self) -> PutAccountDetails<C, M, R>

pub fn put_account_sending_attributes(
&self
) -> PutAccountSendingAttributes<C, M, R>

pub fn put_account_suppression_attributes(
&self
) -> PutAccountSuppressionAttributes<C, M, R>

pub fn put_configuration_set_delivery_options(
&self
) -> PutConfigurationSetDeliveryOptions<C, M, R>

pub fn put_configuration_set_reputation_options(
&self
) -> PutConfigurationSetReputationOptions<C, M, R>

pub fn put_configuration_set_sending_options(
&self
) -> PutConfigurationSetSendingOptions<C, M, R>

pub fn put_configuration_set_suppression_options(
&self
) -> PutConfigurationSetSuppressionOptions<C, M, R>

pub fn put_configuration_set_tracking_options(
&self
) -> PutConfigurationSetTrackingOptions<C, M, R>

pub fn put_dedicated_ip_in_pool(&self) -> PutDedicatedIpInPool<C, M, R>

pub fn put_dedicated_ip_warmup_attributes(
&self
) -> PutDedicatedIpWarmupAttributes<C, M, R>

pub fn put_deliverability_dashboard_option(
&self
) -> PutDeliverabilityDashboardOption<C, M, R>

pub fn put_email_identity_configuration_set_attributes(
&self
) -> PutEmailIdentityConfigurationSetAttributes<C, M, R>

pub fn put_email_identity_dkim_attributes(
&self
) -> PutEmailIdentityDkimAttributes<C, M, R>

pub fn put_email_identity_dkim_signing_attributes(
&self
) -> PutEmailIdentityDkimSigningAttributes<C, M, R>

pub fn put_email_identity_feedback_attributes(
&self
) -> PutEmailIdentityFeedbackAttributes<C, M, R>

pub fn put_email_identity_mail_from_attributes(
&self
) -> PutEmailIdentityMailFromAttributes<C, M, R>

pub fn put_suppressed_destination(&self) -> PutSuppressedDestination<C, M, R>

pub fn send_bulk_email(&self) -> SendBulkEmail<C, M, R>

pub fn send_custom_verification_email(
&self
) -> SendCustomVerificationEmail<C, M, R>

pub fn send_email(&self) -> SendEmail<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn test_render_email_template(&self) -> TestRenderEmailTemplate<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_configuration_set_event_destination(
&self
) -> UpdateConfigurationSetEventDestination<C, M, R>

pub fn update_contact(&self) -> UpdateContact<C, M, R>

pub fn update_contact_list(&self) -> UpdateContactList<C, M, R>

pub fn update_custom_verification_email_template(
&self
) -> UpdateCustomVerificationEmailTemplate<C, M, R>

pub fn update_email_identity_policy(&self) -> UpdateEmailIdentityPolicy<C, M, R>

pub fn update_email_template(&self) -> UpdateEmailTemplate<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self