Module aws_sdk_sesv2::error

Modules

account_suspended_exception

See AccountSuspendedException

already_exists_exception

See AlreadyExistsException

bad_request_exception

See BadRequestException

concurrent_modification_exception

See ConcurrentModificationException

conflict_exception

See ConflictException

invalid_next_token_exception

See InvalidNextTokenException

limit_exceeded_exception

See LimitExceededException

mail_from_domain_not_verified_exception

See MailFromDomainNotVerifiedException

message_rejected

See MessageRejected

not_found_exception

See NotFoundException

sending_paused_exception

See SendingPausedException

too_many_requests_exception

See TooManyRequestsException

Structs

AccountSuspendedException

The message can't be sent because the account's ability to send email has been

AlreadyExistsException

The resource specified in your request already exists.

BadRequestException

The input you provided is invalid.

ConcurrentModificationException

The resource is being modified by another operation or thread.

ConflictException

If there is already an ongoing account details update under review.

CreateConfigurationSetError

CreateConfigurationSetEventDestinationError

CreateContactError

CreateContactListError

CreateCustomVerificationEmailTemplateError

CreateDedicatedIpPoolError

CreateDeliverabilityTestReportError

CreateEmailIdentityError

CreateEmailIdentityPolicyError

CreateEmailTemplateError

CreateImportJobError

DeleteConfigurationSetError

DeleteConfigurationSetEventDestinationError

DeleteContactError

DeleteContactListError

DeleteCustomVerificationEmailTemplateError

DeleteDedicatedIpPoolError

DeleteEmailIdentityError

DeleteEmailIdentityPolicyError

DeleteEmailTemplateError

DeleteSuppressedDestinationError

GetAccountError

GetBlacklistReportsError

GetConfigurationSetError

GetConfigurationSetEventDestinationsError

GetContactError

GetContactListError

GetCustomVerificationEmailTemplateError

GetDedicatedIpError

GetDedicatedIpsError

GetDeliverabilityDashboardOptionsError

GetDeliverabilityTestReportError

GetDomainDeliverabilityCampaignError

GetDomainStatisticsReportError

GetEmailIdentityError

GetEmailIdentityPoliciesError

GetEmailTemplateError

GetImportJobError

GetSuppressedDestinationError

InvalidNextTokenException

The specified request includes an invalid or expired token.

LimitExceededException

There are too many instances of the specified resource type.

ListConfigurationSetsError

ListContactListsError

ListContactsError

ListCustomVerificationEmailTemplatesError

ListDedicatedIpPoolsError

ListDeliverabilityTestReportsError

ListDomainDeliverabilityCampaignsError

ListEmailIdentitiesError

ListEmailTemplatesError

ListImportJobsError

ListSuppressedDestinationsError

ListTagsForResourceError

MailFromDomainNotVerifiedException

The message can't be sent because the sending domain isn't verified.

MessageRejected

The message can't be sent because it contains invalid content.

NotFoundException

The resource you attempted to access doesn't exist.

PutAccountDedicatedIpWarmupAttributesError

PutAccountDetailsError

PutAccountSendingAttributesError

PutAccountSuppressionAttributesError

PutConfigurationSetDeliveryOptionsError

PutConfigurationSetReputationOptionsError

PutConfigurationSetSendingOptionsError

PutConfigurationSetSuppressionOptionsError

PutConfigurationSetTrackingOptionsError

PutDedicatedIpInPoolError

PutDedicatedIpWarmupAttributesError

PutDeliverabilityDashboardOptionError

PutEmailIdentityConfigurationSetAttributesError

PutEmailIdentityDkimAttributesError

PutEmailIdentityDkimSigningAttributesError

PutEmailIdentityFeedbackAttributesError

PutEmailIdentityMailFromAttributesError

PutSuppressedDestinationError

SendBulkEmailError

SendCustomVerificationEmailError

SendEmailError

SendingPausedException

The message can't be sent because the account's ability to send email is currently

TagResourceError

TestRenderEmailTemplateError

TooManyRequestsException

Too many requests have been made to the operation.

UntagResourceError

UpdateConfigurationSetEventDestinationError

UpdateContactError

UpdateContactListError

UpdateCustomVerificationEmailTemplateError

UpdateEmailIdentityPolicyError

UpdateEmailTemplateError

Enums

CreateConfigurationSetErrorKind

CreateConfigurationSetEventDestinationErrorKind

CreateContactErrorKind

CreateContactListErrorKind

CreateCustomVerificationEmailTemplateErrorKind

CreateDedicatedIpPoolErrorKind

CreateDeliverabilityTestReportErrorKind

CreateEmailIdentityErrorKind

CreateEmailIdentityPolicyErrorKind

CreateEmailTemplateErrorKind

CreateImportJobErrorKind

DeleteConfigurationSetErrorKind

DeleteConfigurationSetEventDestinationErrorKind

DeleteContactErrorKind

DeleteContactListErrorKind

DeleteCustomVerificationEmailTemplateErrorKind

DeleteDedicatedIpPoolErrorKind

DeleteEmailIdentityErrorKind

DeleteEmailIdentityPolicyErrorKind

DeleteEmailTemplateErrorKind

DeleteSuppressedDestinationErrorKind

GetAccountErrorKind

GetBlacklistReportsErrorKind

GetConfigurationSetErrorKind

GetConfigurationSetEventDestinationsErrorKind

GetContactErrorKind

GetContactListErrorKind

GetCustomVerificationEmailTemplateErrorKind

GetDedicatedIpErrorKind

GetDedicatedIpsErrorKind

GetDeliverabilityDashboardOptionsErrorKind

GetDeliverabilityTestReportErrorKind

GetDomainDeliverabilityCampaignErrorKind

GetDomainStatisticsReportErrorKind

GetEmailIdentityErrorKind

GetEmailIdentityPoliciesErrorKind

GetEmailTemplateErrorKind

GetImportJobErrorKind

GetSuppressedDestinationErrorKind

ListConfigurationSetsErrorKind

ListContactListsErrorKind

ListContactsErrorKind

ListCustomVerificationEmailTemplatesErrorKind

ListDedicatedIpPoolsErrorKind

ListDeliverabilityTestReportsErrorKind

ListDomainDeliverabilityCampaignsErrorKind

ListEmailIdentitiesErrorKind

ListEmailTemplatesErrorKind

ListImportJobsErrorKind

ListSuppressedDestinationsErrorKind

ListTagsForResourceErrorKind

PutAccountDedicatedIpWarmupAttributesErrorKind

PutAccountDetailsErrorKind

PutAccountSendingAttributesErrorKind

PutAccountSuppressionAttributesErrorKind

PutConfigurationSetDeliveryOptionsErrorKind

PutConfigurationSetReputationOptionsErrorKind

PutConfigurationSetSendingOptionsErrorKind

PutConfigurationSetSuppressionOptionsErrorKind

PutConfigurationSetTrackingOptionsErrorKind

PutDedicatedIpInPoolErrorKind

PutDedicatedIpWarmupAttributesErrorKind

PutDeliverabilityDashboardOptionErrorKind

PutEmailIdentityConfigurationSetAttributesErrorKind

PutEmailIdentityDkimAttributesErrorKind

PutEmailIdentityDkimSigningAttributesErrorKind

PutEmailIdentityFeedbackAttributesErrorKind

PutEmailIdentityMailFromAttributesErrorKind

PutSuppressedDestinationErrorKind

SendBulkEmailErrorKind

SendCustomVerificationEmailErrorKind

SendEmailErrorKind

TagResourceErrorKind

TestRenderEmailTemplateErrorKind

UntagResourceErrorKind

UpdateConfigurationSetEventDestinationErrorKind

UpdateContactErrorKind

UpdateContactListErrorKind

UpdateCustomVerificationEmailTemplateErrorKind

UpdateEmailIdentityPolicyErrorKind

UpdateEmailTemplateErrorKind