Module aws_sdk_sesv2::input

Modules

create_configuration_set_event_destination_input

See CreateConfigurationSetEventDestinationInput

create_configuration_set_input

See CreateConfigurationSetInput

create_contact_input

See CreateContactInput

create_contact_list_input

See CreateContactListInput

create_custom_verification_email_template_input

See CreateCustomVerificationEmailTemplateInput

create_dedicated_ip_pool_input

See CreateDedicatedIpPoolInput

create_deliverability_test_report_input

See CreateDeliverabilityTestReportInput

create_email_identity_input

See CreateEmailIdentityInput

create_email_identity_policy_input

See CreateEmailIdentityPolicyInput

create_email_template_input

See CreateEmailTemplateInput

create_import_job_input

See CreateImportJobInput

delete_configuration_set_event_destination_input

See DeleteConfigurationSetEventDestinationInput

delete_configuration_set_input

See DeleteConfigurationSetInput

delete_contact_input

See DeleteContactInput

delete_contact_list_input

See DeleteContactListInput

delete_custom_verification_email_template_input

See DeleteCustomVerificationEmailTemplateInput

delete_dedicated_ip_pool_input

See DeleteDedicatedIpPoolInput

delete_email_identity_input

See DeleteEmailIdentityInput

delete_email_identity_policy_input

See DeleteEmailIdentityPolicyInput

delete_email_template_input

See DeleteEmailTemplateInput

delete_suppressed_destination_input

See DeleteSuppressedDestinationInput

get_account_input

See GetAccountInput

get_blacklist_reports_input

See GetBlacklistReportsInput

get_configuration_set_event_destinations_input

See GetConfigurationSetEventDestinationsInput

get_configuration_set_input

See GetConfigurationSetInput

get_contact_input

See GetContactInput

get_contact_list_input

See GetContactListInput

get_custom_verification_email_template_input

See GetCustomVerificationEmailTemplateInput

get_dedicated_ip_input

See GetDedicatedIpInput

get_dedicated_ips_input

See GetDedicatedIpsInput

get_deliverability_dashboard_options_input

See GetDeliverabilityDashboardOptionsInput

get_deliverability_test_report_input

See GetDeliverabilityTestReportInput

get_domain_deliverability_campaign_input

See GetDomainDeliverabilityCampaignInput

get_domain_statistics_report_input

See GetDomainStatisticsReportInput

get_email_identity_input

See GetEmailIdentityInput

get_email_identity_policies_input

See GetEmailIdentityPoliciesInput

get_email_template_input

See GetEmailTemplateInput

get_import_job_input

See GetImportJobInput

get_suppressed_destination_input

See GetSuppressedDestinationInput

list_configuration_sets_input

See ListConfigurationSetsInput

list_contact_lists_input

See ListContactListsInput

list_contacts_input

See ListContactsInput

list_custom_verification_email_templates_input

See ListCustomVerificationEmailTemplatesInput

list_dedicated_ip_pools_input

See ListDedicatedIpPoolsInput

list_deliverability_test_reports_input

See ListDeliverabilityTestReportsInput

list_domain_deliverability_campaigns_input

See ListDomainDeliverabilityCampaignsInput

list_email_identities_input

See ListEmailIdentitiesInput

list_email_templates_input

See ListEmailTemplatesInput

list_import_jobs_input

See ListImportJobsInput

list_suppressed_destinations_input

See ListSuppressedDestinationsInput

list_tags_for_resource_input

See ListTagsForResourceInput

put_account_dedicated_ip_warmup_attributes_input

See PutAccountDedicatedIpWarmupAttributesInput

put_account_details_input

See PutAccountDetailsInput

put_account_sending_attributes_input

See PutAccountSendingAttributesInput

put_account_suppression_attributes_input

See PutAccountSuppressionAttributesInput

put_configuration_set_delivery_options_input

See PutConfigurationSetDeliveryOptionsInput

put_configuration_set_reputation_options_input

See PutConfigurationSetReputationOptionsInput

put_configuration_set_sending_options_input

See PutConfigurationSetSendingOptionsInput

put_configuration_set_suppression_options_input

See PutConfigurationSetSuppressionOptionsInput

put_configuration_set_tracking_options_input

See PutConfigurationSetTrackingOptionsInput

put_dedicated_ip_in_pool_input

See PutDedicatedIpInPoolInput

put_dedicated_ip_warmup_attributes_input

See PutDedicatedIpWarmupAttributesInput

put_deliverability_dashboard_option_input

See PutDeliverabilityDashboardOptionInput

put_email_identity_configuration_set_attributes_input

See PutEmailIdentityConfigurationSetAttributesInput

put_email_identity_dkim_attributes_input

See PutEmailIdentityDkimAttributesInput

put_email_identity_dkim_signing_attributes_input

See PutEmailIdentityDkimSigningAttributesInput

put_email_identity_feedback_attributes_input

See PutEmailIdentityFeedbackAttributesInput

put_email_identity_mail_from_attributes_input

See PutEmailIdentityMailFromAttributesInput

put_suppressed_destination_input

See PutSuppressedDestinationInput

send_bulk_email_input

See SendBulkEmailInput

send_custom_verification_email_input

See SendCustomVerificationEmailInput

send_email_input

See SendEmailInput

tag_resource_input

See TagResourceInput

test_render_email_template_input

See TestRenderEmailTemplateInput

untag_resource_input

See UntagResourceInput

update_configuration_set_event_destination_input

See UpdateConfigurationSetEventDestinationInput

update_contact_input

See UpdateContactInput

update_contact_list_input

See UpdateContactListInput

update_custom_verification_email_template_input

See UpdateCustomVerificationEmailTemplateInput

update_email_identity_policy_input

See UpdateEmailIdentityPolicyInput

update_email_template_input

See UpdateEmailTemplateInput

Structs

CreateConfigurationSetEventDestinationInput

A request to add an event destination to a configuration set.

CreateConfigurationSetInput

A request to create a configuration set.

CreateContactInput

CreateContactListInput

CreateCustomVerificationEmailTemplateInput

Represents a request to create a custom verification email template.

CreateDedicatedIpPoolInput

A request to create a new dedicated IP pool.

CreateDeliverabilityTestReportInput

A request to perform a predictive inbox placement test. Predictive inbox placement tests can help you predict how your messages will

CreateEmailIdentityInput

A request to begin the verification process for an email identity (an email address or

CreateEmailIdentityPolicyInput

Represents a request to create a sending authorization policy for an identity. Sending

CreateEmailTemplateInput

Represents a request to create an email template. For more information, see the Amazon SES

CreateImportJobInput

Represents a request to create an import job from a data source for a data

DeleteConfigurationSetEventDestinationInput

A request to delete an event destination from a configuration set.

DeleteConfigurationSetInput

A request to delete a configuration set.

DeleteContactInput

DeleteContactListInput

DeleteCustomVerificationEmailTemplateInput

Represents a request to delete an existing custom verification email template.

DeleteDedicatedIpPoolInput

A request to delete a dedicated IP pool.

DeleteEmailIdentityInput

A request to delete an existing email identity. When you delete an identity, you lose

DeleteEmailIdentityPolicyInput

Represents a request to delete a sending authorization policy for an identity. Sending

DeleteEmailTemplateInput

Represents a request to delete an email template. For more information, see the Amazon SES Developer

DeleteSuppressedDestinationInput

A request to remove an email address from the suppression list for your

GetAccountInput

A request to obtain information about the email-sending capabilities of your Amazon SES

GetBlacklistReportsInput

A request to retrieve a list of the blacklists that your dedicated IP addresses appear

GetConfigurationSetEventDestinationsInput

A request to obtain information about the event destinations for a configuration

GetConfigurationSetInput

A request to obtain information about a configuration set.

GetContactInput

GetContactListInput

GetCustomVerificationEmailTemplateInput

Represents a request to retrieve an existing custom verification email

GetDedicatedIpInput

A request to obtain more information about a dedicated IP address.

GetDedicatedIpsInput

A request to obtain more information about dedicated IP pools.

GetDeliverabilityDashboardOptionsInput

Retrieve information about the status of the Deliverability dashboard for your AWS account.

GetDeliverabilityTestReportInput

A request to retrieve the results of a predictive inbox placement test.

GetDomainDeliverabilityCampaignInput

Retrieve all the deliverability data for a specific campaign. This data is available

GetDomainStatisticsReportInput

A request to obtain deliverability metrics for a domain.

GetEmailIdentityInput

A request to return details about an email identity.

GetEmailIdentityPoliciesInput

A request to return the policies of an email identity.

GetEmailTemplateInput

Represents a request to display the template object (which includes the subject line,

GetImportJobInput

Represents a request for information about an import job using the import job

GetSuppressedDestinationInput

A request to retrieve information about an email address that's on the suppression

ListConfigurationSetsInput

A request to obtain a list of configuration sets for your Amazon SES account in the current

ListContactListsInput

ListContactsInput

ListCustomVerificationEmailTemplatesInput

Represents a request to list the existing custom verification email templates for your

ListDedicatedIpPoolsInput

A request to obtain a list of dedicated IP pools.

ListDeliverabilityTestReportsInput

A request to list all of the predictive inbox placement tests that you've performed.

ListDomainDeliverabilityCampaignsInput

Retrieve deliverability data for all the campaigns that used a specific domain to send

ListEmailIdentitiesInput

A request to list all of the email identities associated with your AWS account. This

ListEmailTemplatesInput

Represents a request to list the email templates present in your Amazon SES account in the

ListImportJobsInput

Represents a request to list all of the import jobs for a data destination within the

ListSuppressedDestinationsInput

A request to obtain a list of email destinations that are on the suppression list for

ListTagsForResourceInput

PutAccountDedicatedIpWarmupAttributesInput

A request to enable or disable the automatic IP address warm-up feature.

PutAccountDetailsInput

A request to submit new account details.

PutAccountSendingAttributesInput

A request to change the ability of your account to send email.

PutAccountSuppressionAttributesInput

A request to change your account's suppression preferences.

PutConfigurationSetDeliveryOptionsInput

A request to associate a configuration set with a dedicated IP pool.

PutConfigurationSetReputationOptionsInput

A request to enable or disable tracking of reputation metrics for a configuration

PutConfigurationSetSendingOptionsInput

A request to enable or disable the ability of Amazon SES to send emails that use a specific

PutConfigurationSetSuppressionOptionsInput

A request to change the account suppression list preferences for a specific

PutConfigurationSetTrackingOptionsInput

A request to add a custom domain for tracking open and click events to a configuration

PutDedicatedIpInPoolInput

A request to move a dedicated IP address to a dedicated IP pool.

PutDedicatedIpWarmupAttributesInput

A request to change the warm-up attributes for a dedicated IP address. This operation

PutDeliverabilityDashboardOptionInput

Enable or disable the Deliverability dashboard. When you enable the Deliverability dashboard, you gain

PutEmailIdentityConfigurationSetAttributesInput

A request to associate a configuration set with an email identity.

PutEmailIdentityDkimAttributesInput

A request to enable or disable DKIM signing of email that you send from an email

PutEmailIdentityDkimSigningAttributesInput

A request to change the DKIM attributes for an email identity.

PutEmailIdentityFeedbackAttributesInput

A request to set the attributes that control how bounce and complaint events are

PutEmailIdentityMailFromAttributesInput

A request to configure the custom MAIL FROM domain for a verified identity.

PutSuppressedDestinationInput

A request to add an email destination to the suppression list for your account.

SendBulkEmailInput

Represents a request to send email messages to multiple destinations using Amazon SES. For

SendCustomVerificationEmailInput

Represents a request to send a custom verification email to a specified

SendEmailInput

Represents a request to send a single formatted email using Amazon SES. For more

TagResourceInput

TestRenderEmailTemplateInput

>Represents a request to create a preview of the MIME content of an email when

UntagResourceInput

UpdateConfigurationSetEventDestinationInput

A request to change the settings for an event destination for a configuration

UpdateContactInput

UpdateContactListInput

UpdateCustomVerificationEmailTemplateInput

Represents a request to update an existing custom verification email template.

UpdateEmailIdentityPolicyInput

Represents a request to update a sending authorization policy for an identity. Sending

UpdateEmailTemplateInput

Represents a request to update an email template. For more information, see the Amazon SES