Module aws_sdk_sesv2::model

Modules

account_details

See AccountDetails

blacklist_entry

See BlacklistEntry

body

See Body

bulk_email_content

See BulkEmailContent

bulk_email_entry

See BulkEmailEntry

bulk_email_entry_result

See BulkEmailEntryResult

cloud_watch_destination

See CloudWatchDestination

cloud_watch_dimension_configuration

See CloudWatchDimensionConfiguration

contact

See Contact

contact_list

See ContactList

contact_list_destination

See ContactListDestination

content

See Content

custom_verification_email_template_metadata

See CustomVerificationEmailTemplateMetadata

daily_volume

See DailyVolume

dedicated_ip

See DedicatedIp

deliverability_test_report

See DeliverabilityTestReport

delivery_options

See DeliveryOptions

destination

See Destination

dkim_attributes

See DkimAttributes

dkim_signing_attributes

See DkimSigningAttributes

domain_deliverability_campaign

See DomainDeliverabilityCampaign

domain_deliverability_tracking_option

See DomainDeliverabilityTrackingOption

domain_isp_placement

See DomainIspPlacement

email_content

See EmailContent

email_template_content

See EmailTemplateContent

email_template_metadata

See EmailTemplateMetadata

event_destination

See EventDestination

event_destination_definition

See EventDestinationDefinition

failure_info

See FailureInfo

identity_info

See IdentityInfo

import_data_source

See ImportDataSource

import_destination

See ImportDestination

import_job_summary

See ImportJobSummary

inbox_placement_tracking_option

See InboxPlacementTrackingOption

isp_placement

See IspPlacement

kinesis_firehose_destination

See KinesisFirehoseDestination

list_contacts_filter

See ListContactsFilter

list_management_options

See ListManagementOptions

mail_from_attributes

See MailFromAttributes

message

See Message

message_tag

See MessageTag

overall_volume

See OverallVolume

pinpoint_destination

See PinpointDestination

placement_statistics

See PlacementStatistics

raw_message

See RawMessage

replacement_email_content

See ReplacementEmailContent

replacement_template

See ReplacementTemplate

reputation_options

See ReputationOptions

review_details

See ReviewDetails

send_quota

See SendQuota

sending_options

See SendingOptions

sns_destination

See SnsDestination

suppressed_destination

See SuppressedDestination

suppressed_destination_attributes

See SuppressedDestinationAttributes

suppressed_destination_summary

See SuppressedDestinationSummary

suppression_attributes

See SuppressionAttributes

suppression_list_destination

See SuppressionListDestination

suppression_options

See SuppressionOptions

tag

See Tag

template

See Template

topic

See Topic

topic_filter

See TopicFilter

topic_preference

See TopicPreference

tracking_options

See TrackingOptions

volume_statistics

See VolumeStatistics

Structs

AccountDetails

An object that contains information about your account details.

BlacklistEntry

An object that contains information about a blacklisting event that impacts one of the

Body

Represents the body of the email message.

BulkEmailContent

An object that contains the body of the message. You can specify a template

BulkEmailEntry

BulkEmailEntryResult

The result of the SendBulkEmail operation of each specified

CloudWatchDestination

An object that defines an Amazon CloudWatch destination for email events. You can use Amazon CloudWatch to

CloudWatchDimensionConfiguration

An object that defines the dimension configuration to use when you send email events

Contact

A contact is the end-user who is receiving the email.

ContactList

A list that contains contacts that have subscribed to a particular topic or

ContactListDestination

An object that contains details about the action of a contact list.

Content

An object that represents the content of the email, and optionally a character set

CustomVerificationEmailTemplateMetadata

Contains information about a custom verification email template.

DailyVolume

An object that contains information about the volume of email sent on each day of the

DedicatedIp

Contains information about a dedicated IP address that is associated with your Amazon SES

DeliverabilityTestReport

An object that contains metadata related to a predictive inbox placement test.

DeliveryOptions

Used to associate a configuration set with a dedicated IP pool.

Destination

An object that describes the recipients for an email.

DkimAttributes

An object that contains information about the DKIM authentication status for an email

DkimSigningAttributes

An object that contains information about the tokens used for setting up Bring Your

DomainDeliverabilityCampaign

An object that contains the deliverability data for a specific campaign. This data is

DomainDeliverabilityTrackingOption

An object that contains information about the Deliverability dashboard subscription for a

DomainIspPlacement

An object that contains inbox placement data for email sent from one of your email

EmailContent

An object that defines the entire content of the email, including the message headers

EmailTemplateContent

The content of the email, composed of a subject line, an HTML part, and a text-only

EmailTemplateMetadata

Contains information about an email template.

EventDestination

In the Amazon SES API v2, events include message sends, deliveries, opens,

EventDestinationDefinition

An object that defines the event destination. Specifically, it defines which services

FailureInfo

An object that contains the failure details about an import job.

IdentityInfo

Information about an email identity.

ImportDataSource

An object that contains details about the data source of the import job.

ImportDestination

An object that contains details about the resource destination the import job is going

ImportJobSummary

A summary of the import job.

InboxPlacementTrackingOption

An object that contains information about the inbox placement data settings for a

IspPlacement

An object that describes how email sent during the predictive inbox placement test was handled by a certain

KinesisFirehoseDestination

An object that defines an Amazon Kinesis Data Firehose destination for email events. You can use Amazon Kinesis Data Firehose to

ListContactsFilter

A filter that can be applied to a list of contacts.

ListManagementOptions

An object used to specify a list or topic to which an email belongs, which will be

MailFromAttributes

A list of attributes that are associated with a MAIL FROM domain.

Message

Represents the email message that you're sending. The Message object

MessageTag

Contains the name and value of a tag that you apply to an email. You can use message

OverallVolume

An object that contains information about email that was sent from the selected

PinpointDestination

An object that defines an Amazon Pinpoint project destination for email events. You can send

PlacementStatistics

An object that contains inbox placement data for an email provider.

RawMessage

Represents the raw content of an email message.

ReplacementEmailContent

The ReplaceEmailContent object to be used for a specific

ReplacementTemplate

An object which contains ReplacementTemplateData to be used for a

ReputationOptions

Enable or disable collection of reputation metrics for emails that you send using this

ReviewDetails

An object that contains information about your account details review.

SendQuota

An object that contains information about the per-day and per-second sending limits

SendingOptions

Used to enable or disable email sending for messages that use this configuration set

SnsDestination

An object that defines an Amazon SNS destination for email events. You can use Amazon SNS to

SuppressedDestination

An object that contains information about an email address that is on the suppression

SuppressedDestinationAttributes

An object that contains additional attributes that are related an email address that

SuppressedDestinationSummary

A summary that describes the suppressed email address.

SuppressionAttributes

An object that contains information about the email address suppression preferences

SuppressionListDestination

An object that contains details about the action of suppression list.

SuppressionOptions

An object that contains information about the suppression list preferences for your

Tag

An object that defines the tags that are associated with a resource.

Template

An object that defines the email template to use for an email message, and the values

Topic

An interest group, theme, or label within a list. Lists can have multiple

TopicFilter

Used for filtering by a specific topic preference.

TopicPreference

The contact's preference for being opted-in to or opted-out of a topic.

TrackingOptions

An object that defines the tracking options for a configuration set. When you use the

VolumeStatistics

An object that contains information about the amount of email that was delivered to

Enums

BehaviorOnMxFailure

The action that you want to take if the required MX record can't be found when you

BulkEmailStatus

ContactLanguage

ContactListImportAction

DataFormat

The data format of the import job's data source.

DeliverabilityDashboardAccountStatus

The current status of your Deliverability dashboard subscription. If this value is

DeliverabilityTestStatus

The status of a predictive inbox placement test. If the status is IN_PROGRESS, then the predictive inbox placement test is

DimensionValueSource

The location where the Amazon SES API v2 finds the value of a dimension to publish to Amazon CloudWatch. If

DkimSigningAttributesOrigin

DkimStatus

The DKIM authentication status of the identity. The status can be one of the

EventType

An email sending event type. For example, email sends, opens, and bounces are all

IdentityType

The email identity type. The identity type can be one of the following:

ImportDestinationType

The destination of the import job, which can be used to list import jobs that have a

JobStatus

The status of the import job.

MailFromDomainStatus

The status of the MAIL FROM domain. This status can have the following values:

MailType

ReviewStatus

SubscriptionStatus

SuppressionListImportAction

The type of action that you want to perform on the address. Acceptable values:

SuppressionListReason

The reason that the address was added to the suppression list for your account. The

TlsPolicy

Specifies whether messages that use the configuration set are required to use

WarmupStatus

The warmup status of a dedicated IP.