Module aws_sdk_sesv2::operation

Structs

CreateConfigurationSet

Create a configuration set. Configuration sets are groups of

CreateConfigurationSetEventDestination

Create an event destination. Events include message sends,

CreateContact

Creates a contact, which is an end-user who is receiving the email, and adds them to a

CreateContactList

Creates a contact list.

CreateCustomVerificationEmailTemplate

Creates a new custom verification email template.

CreateDedicatedIpPool

Create a new pool of dedicated IP addresses. A pool can include one or more dedicated

CreateDeliverabilityTestReport

Create a new predictive inbox placement test. Predictive inbox placement tests can help you predict how your messages will be handled

CreateEmailIdentity

Starts the process of verifying an email identity. An identity is

CreateEmailIdentityPolicy

Creates the specified sending authorization policy for the given identity (an email

CreateEmailTemplate

Creates an email template. Email templates enable you to send personalized email to

CreateImportJob

Creates an import job for a data destination.

DeleteConfigurationSet

Delete an existing configuration set.

DeleteConfigurationSetEventDestination

Delete an event destination.

DeleteContact

Removes a contact from a contact list.

DeleteContactList

Deletes a contact list and all of the contacts on that list.

DeleteCustomVerificationEmailTemplate

Deletes an existing custom verification email template.

DeleteDedicatedIpPool

Delete a dedicated IP pool.

DeleteEmailIdentity

Deletes an email identity. An identity can be either an email address or a domain

DeleteEmailIdentityPolicy

Deletes the specified sending authorization policy for the given identity (an email

DeleteEmailTemplate

Deletes an email template.

DeleteSuppressedDestination

Removes an email address from the suppression list for your account.

GetAccount

Obtain information about the email-sending status and capabilities of your Amazon SES

GetBlacklistReports

Retrieve a list of the blacklists that your dedicated IP addresses appear on.

GetConfigurationSet

Get information about an existing configuration set, including the dedicated IP pool

GetConfigurationSetEventDestinations

Retrieve a list of event destinations that are associated with a configuration

GetContact

Returns a contact from a contact list.

GetContactList

Returns contact list metadata. It does not return any information about the contacts

GetCustomVerificationEmailTemplate

Returns the custom email verification template for the template name you

GetDedicatedIp

Get information about a dedicated IP address, including the name of the dedicated IP

GetDedicatedIps

List the dedicated IP addresses that are associated with your AWS

GetDeliverabilityDashboardOptions

Retrieve information about the status of the Deliverability dashboard for your account. When

GetDeliverabilityTestReport

Retrieve the results of a predictive inbox placement test.

GetDomainDeliverabilityCampaign

Retrieve all the deliverability data for a specific campaign. This data is available

GetDomainStatisticsReport

Retrieve inbox placement and engagement rates for the domains that you use to send

GetEmailIdentity

Provides information about a specific identity, including the identity's verification

GetEmailIdentityPolicies

Returns the requested sending authorization policies for the given identity (an email

GetEmailTemplate

Displays the template object (which includes the subject line, HTML part and text

GetImportJob

Provides information about an import job.

GetSuppressedDestination

Retrieves information about a specific email address that's on the suppression list

ListConfigurationSets

List all of the configuration sets associated with your account in the current

ListContactLists

Lists all of the contact lists available.

ListContacts

Lists the contacts present in a specific contact list.

ListCustomVerificationEmailTemplates

Lists the existing custom verification email templates for your account in the current

ListDedicatedIpPools

List all of the dedicated IP pools that exist in your AWS account in the current

ListDeliverabilityTestReports

Show a list of the predictive inbox placement tests that you've performed, regardless of their statuses. For

ListDomainDeliverabilityCampaigns

Retrieve deliverability data for all the campaigns that used a specific domain to send

ListEmailIdentities

Returns a list of all of the email identities that are associated with your AWS

ListEmailTemplates

Lists the email templates present in your Amazon SES account in the current AWS

ListImportJobs

Lists all of the import jobs.

ListSuppressedDestinations

Retrieves a list of email addresses that are on the suppression list for your

ListTagsForResource

Retrieve a list of the tags (keys and values) that are associated with a specified

PutAccountDedicatedIpWarmupAttributes

Enable or disable the automatic warm-up feature for dedicated IP addresses.

PutAccountDetails

Update your Amazon SES account details.

PutAccountSendingAttributes

Enable or disable the ability of your account to send email.

PutAccountSuppressionAttributes

Change the settings for the account-level suppression list.

PutConfigurationSetDeliveryOptions

Associate a configuration set with a dedicated IP pool. You can use dedicated IP pools

PutConfigurationSetReputationOptions

Enable or disable collection of reputation metrics for emails that you send using a

PutConfigurationSetSendingOptions

Enable or disable email sending for messages that use a particular configuration set

PutConfigurationSetSuppressionOptions

Specify the account suppression list preferences for a configuration set.

PutConfigurationSetTrackingOptions

Specify a custom domain to use for open and click tracking elements in email that you

PutDedicatedIpInPool

Move a dedicated IP address to an existing dedicated IP pool.

PutDedicatedIpWarmupAttributes

PutDeliverabilityDashboardOption

Enable or disable the Deliverability dashboard. When you enable the Deliverability dashboard, you gain

PutEmailIdentityConfigurationSetAttributes

Used to associate a configuration set with an email identity.

PutEmailIdentityDkimAttributes

Used to enable or disable DKIM authentication for an email identity.

PutEmailIdentityDkimSigningAttributes

Used to configure or change the DKIM authentication settings for an email domain

PutEmailIdentityFeedbackAttributes

Used to enable or disable feedback forwarding for an identity. This setting determines

PutEmailIdentityMailFromAttributes

Used to enable or disable the custom Mail-From domain configuration for an email

PutSuppressedDestination

Adds an email address to the suppression list for your account.

SendBulkEmail

Composes an email message to multiple destinations.

SendCustomVerificationEmail

Adds an email address to the list of identities for your Amazon SES account in the current

SendEmail

Sends an email message. You can use the Amazon SES API v2 to send two types of

TagResource

Add one or more tags (keys and values) to a specified resource. A

TestRenderEmailTemplate

Creates a preview of the MIME content of an email when provided with a template and a

UntagResource

Remove one or more tags (keys and values) from a specified resource.

UpdateConfigurationSetEventDestination

Update the configuration of an event destination for a configuration set.

UpdateContact

Updates a contact's preferences for a list. It is not necessary to specify all

UpdateContactList

Updates contact list metadata. This operation does a complete replacement.

UpdateCustomVerificationEmailTemplate

Updates an existing custom verification email template.

UpdateEmailIdentityPolicy

Updates the specified sending authorization policy for the given identity (an email

UpdateEmailTemplate

Updates an email template. Email templates enable you to send personalized email to