Module aws_sdk_sesv2::output

Modules

create_configuration_set_event_destination_output

See CreateConfigurationSetEventDestinationOutput

create_configuration_set_output

See CreateConfigurationSetOutput

create_contact_list_output

See CreateContactListOutput

create_contact_output

See CreateContactOutput

create_custom_verification_email_template_output

See CreateCustomVerificationEmailTemplateOutput

create_dedicated_ip_pool_output

See CreateDedicatedIpPoolOutput

create_deliverability_test_report_output

See CreateDeliverabilityTestReportOutput

create_email_identity_output

See CreateEmailIdentityOutput

create_email_identity_policy_output

See CreateEmailIdentityPolicyOutput

create_email_template_output

See CreateEmailTemplateOutput

create_import_job_output

See CreateImportJobOutput

delete_configuration_set_event_destination_output

See DeleteConfigurationSetEventDestinationOutput

delete_configuration_set_output

See DeleteConfigurationSetOutput

delete_contact_list_output

See DeleteContactListOutput

delete_contact_output

See DeleteContactOutput

delete_custom_verification_email_template_output

See DeleteCustomVerificationEmailTemplateOutput

delete_dedicated_ip_pool_output

See DeleteDedicatedIpPoolOutput

delete_email_identity_output

See DeleteEmailIdentityOutput

delete_email_identity_policy_output

See DeleteEmailIdentityPolicyOutput

delete_email_template_output

See DeleteEmailTemplateOutput

delete_suppressed_destination_output

See DeleteSuppressedDestinationOutput

get_account_output

See GetAccountOutput

get_blacklist_reports_output

See GetBlacklistReportsOutput

get_configuration_set_event_destinations_output

See GetConfigurationSetEventDestinationsOutput

get_configuration_set_output

See GetConfigurationSetOutput

get_contact_list_output

See GetContactListOutput

get_contact_output

See GetContactOutput

get_custom_verification_email_template_output

See GetCustomVerificationEmailTemplateOutput

get_dedicated_ip_output

See GetDedicatedIpOutput

get_dedicated_ips_output

See GetDedicatedIpsOutput

get_deliverability_dashboard_options_output

See GetDeliverabilityDashboardOptionsOutput

get_deliverability_test_report_output

See GetDeliverabilityTestReportOutput

get_domain_deliverability_campaign_output

See GetDomainDeliverabilityCampaignOutput

get_domain_statistics_report_output

See GetDomainStatisticsReportOutput

get_email_identity_output

See GetEmailIdentityOutput

get_email_identity_policies_output

See GetEmailIdentityPoliciesOutput

get_email_template_output

See GetEmailTemplateOutput

get_import_job_output

See GetImportJobOutput

get_suppressed_destination_output

See GetSuppressedDestinationOutput

list_configuration_sets_output

See ListConfigurationSetsOutput

list_contact_lists_output

See ListContactListsOutput

list_contacts_output

See ListContactsOutput

list_custom_verification_email_templates_output

See ListCustomVerificationEmailTemplatesOutput

list_dedicated_ip_pools_output

See ListDedicatedIpPoolsOutput

list_deliverability_test_reports_output

See ListDeliverabilityTestReportsOutput

list_domain_deliverability_campaigns_output

See ListDomainDeliverabilityCampaignsOutput

list_email_identities_output

See ListEmailIdentitiesOutput

list_email_templates_output

See ListEmailTemplatesOutput

list_import_jobs_output

See ListImportJobsOutput

list_suppressed_destinations_output

See ListSuppressedDestinationsOutput

list_tags_for_resource_output

See ListTagsForResourceOutput

put_account_dedicated_ip_warmup_attributes_output

See PutAccountDedicatedIpWarmupAttributesOutput

put_account_details_output

See PutAccountDetailsOutput

put_account_sending_attributes_output

See PutAccountSendingAttributesOutput

put_account_suppression_attributes_output

See PutAccountSuppressionAttributesOutput

put_configuration_set_delivery_options_output

See PutConfigurationSetDeliveryOptionsOutput

put_configuration_set_reputation_options_output

See PutConfigurationSetReputationOptionsOutput

put_configuration_set_sending_options_output

See PutConfigurationSetSendingOptionsOutput

put_configuration_set_suppression_options_output

See PutConfigurationSetSuppressionOptionsOutput

put_configuration_set_tracking_options_output

See PutConfigurationSetTrackingOptionsOutput

put_dedicated_ip_in_pool_output

See PutDedicatedIpInPoolOutput

put_dedicated_ip_warmup_attributes_output

See PutDedicatedIpWarmupAttributesOutput

put_deliverability_dashboard_option_output

See PutDeliverabilityDashboardOptionOutput

put_email_identity_configuration_set_attributes_output

See PutEmailIdentityConfigurationSetAttributesOutput

put_email_identity_dkim_attributes_output

See PutEmailIdentityDkimAttributesOutput

put_email_identity_dkim_signing_attributes_output

See PutEmailIdentityDkimSigningAttributesOutput

put_email_identity_feedback_attributes_output

See PutEmailIdentityFeedbackAttributesOutput

put_email_identity_mail_from_attributes_output

See PutEmailIdentityMailFromAttributesOutput

put_suppressed_destination_output

See PutSuppressedDestinationOutput

send_bulk_email_output

See SendBulkEmailOutput

send_custom_verification_email_output

See SendCustomVerificationEmailOutput

send_email_output

See SendEmailOutput

tag_resource_output

See TagResourceOutput

test_render_email_template_output

See TestRenderEmailTemplateOutput

untag_resource_output

See UntagResourceOutput

update_configuration_set_event_destination_output

See UpdateConfigurationSetEventDestinationOutput

update_contact_list_output

See UpdateContactListOutput

update_contact_output

See UpdateContactOutput

update_custom_verification_email_template_output

See UpdateCustomVerificationEmailTemplateOutput

update_email_identity_policy_output

See UpdateEmailIdentityPolicyOutput

update_email_template_output

See UpdateEmailTemplateOutput

Structs

CreateConfigurationSetEventDestinationOutput

An HTTP 200 response if the request succeeds, or an error message if the request

CreateConfigurationSetOutput

An HTTP 200 response if the request succeeds, or an error message if the request

CreateContactListOutput

CreateContactOutput

CreateCustomVerificationEmailTemplateOutput

If the action is successful, the service sends back an HTTP 200 response with an empty

CreateDedicatedIpPoolOutput

An HTTP 200 response if the request succeeds, or an error message if the request

CreateDeliverabilityTestReportOutput

Information about the predictive inbox placement test that you created.

CreateEmailIdentityOutput

If the email identity is a domain, this object contains information about the DKIM

CreateEmailIdentityPolicyOutput

An HTTP 200 response if the request succeeds, or an error message if the request

CreateEmailTemplateOutput

If the action is successful, the service sends back an HTTP 200 response with an empty

CreateImportJobOutput

An HTTP 200 response if the request succeeds, or an error message if the request

DeleteConfigurationSetEventDestinationOutput

An HTTP 200 response if the request succeeds, or an error message if the request

DeleteConfigurationSetOutput

An HTTP 200 response if the request succeeds, or an error message if the request

DeleteContactListOutput

DeleteContactOutput

DeleteCustomVerificationEmailTemplateOutput

If the action is successful, the service sends back an HTTP 200 response with an empty

DeleteDedicatedIpPoolOutput

An HTTP 200 response if the request succeeds, or an error message if the request

DeleteEmailIdentityOutput

An HTTP 200 response if the request succeeds, or an error message if the request

DeleteEmailIdentityPolicyOutput

An HTTP 200 response if the request succeeds, or an error message if the request

DeleteEmailTemplateOutput

If the action is successful, the service sends back an HTTP 200 response with an empty

DeleteSuppressedDestinationOutput

An HTTP 200 response if the request succeeds, or an error message if the request

GetAccountOutput

A list of details about the email-sending capabilities of your Amazon SES account in the

GetBlacklistReportsOutput

An object that contains information about blacklist events.

GetConfigurationSetEventDestinationsOutput

Information about an event destination for a configuration set.

GetConfigurationSetOutput

Information about a configuration set.

GetContactListOutput

GetContactOutput

GetCustomVerificationEmailTemplateOutput

The following elements are returned by the service.

GetDedicatedIpOutput

Information about a dedicated IP address.

GetDedicatedIpsOutput

Information about the dedicated IP addresses that are associated with your AWS

GetDeliverabilityDashboardOptionsOutput

An object that shows the status of the Deliverability dashboard.

GetDeliverabilityTestReportOutput

The results of the predictive inbox placement test.

GetDomainDeliverabilityCampaignOutput

An object that contains all the deliverability data for a specific campaign. This data

GetDomainStatisticsReportOutput

An object that includes statistics that are related to the domain that you

GetEmailIdentityOutput

Details about an email identity.

GetEmailIdentityPoliciesOutput

Identity policies associated with email identity.

GetEmailTemplateOutput

The following element is returned by the service.

GetImportJobOutput

An HTTP 200 response if the request succeeds, or an error message if the request

GetSuppressedDestinationOutput

Information about the suppressed email address.

ListConfigurationSetsOutput

A list of configuration sets in your Amazon SES account in the current AWS Region.

ListContactListsOutput

ListContactsOutput

ListCustomVerificationEmailTemplatesOutput

The following elements are returned by the service.

ListDedicatedIpPoolsOutput

A list of dedicated IP pools.

ListDeliverabilityTestReportsOutput

A list of the predictive inbox placement test reports that are available for your account, regardless of

ListDomainDeliverabilityCampaignsOutput

An array of objects that provide deliverability data for all the campaigns that used a

ListEmailIdentitiesOutput

A list of all of the identities that you've attempted to verify, regardless of whether

ListEmailTemplatesOutput

The following elements are returned by the service.

ListImportJobsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

ListSuppressedDestinationsOutput

A list of suppressed email addresses.

ListTagsForResourceOutput

PutAccountDedicatedIpWarmupAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutAccountDetailsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutAccountSendingAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutAccountSuppressionAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutConfigurationSetDeliveryOptionsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutConfigurationSetReputationOptionsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutConfigurationSetSendingOptionsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutConfigurationSetSuppressionOptionsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutConfigurationSetTrackingOptionsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutDedicatedIpInPoolOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutDedicatedIpWarmupAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutDeliverabilityDashboardOptionOutput

A response that indicates whether the Deliverability dashboard is enabled.

PutEmailIdentityConfigurationSetAttributesOutput

If the action is successful, the service sends back an HTTP 200 response with an empty HTTP body.

PutEmailIdentityDkimAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutEmailIdentityDkimSigningAttributesOutput

If the action is successful, the service sends back an HTTP 200 response.

PutEmailIdentityFeedbackAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutEmailIdentityMailFromAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutSuppressedDestinationOutput

An HTTP 200 response if the request succeeds, or an error message if the request

SendBulkEmailOutput

The following data is returned in JSON format by the service.

SendCustomVerificationEmailOutput

The following element is returned by the service.

SendEmailOutput

A unique message ID that you receive when an email is accepted for sending.

TagResourceOutput

TestRenderEmailTemplateOutput

The following element is returned by the service.

UntagResourceOutput

UpdateConfigurationSetEventDestinationOutput

An HTTP 200 response if the request succeeds, or an error message if the request

UpdateContactListOutput

UpdateContactOutput

UpdateCustomVerificationEmailTemplateOutput

If the action is successful, the service sends back an HTTP 200 response with an empty

UpdateEmailIdentityPolicyOutput

An HTTP 200 response if the request succeeds, or an error message if the request

UpdateEmailTemplateOutput

If the action is successful, the service sends back an HTTP 200 response with an empty